add toolTip
