+++ src/components/index.tsx
@@ -5,6 +5,7 @@ export { default as ConfirmationModal } from './ConfirmationModal';
 export { default as Root } from './Root';
 export { default as Select } from './Select';
 export { default as Tabs } from './Tabs';
-export { default as Tooltip } from './Tooltip';
+export { default as TableTooltip } from './TableTooltip';
 export { default as ButtonGroup } from './ButtonGroup';
 export type { ButtonGroupType } from './ButtonGroup';
+export { default as Tooltip } from './Tooltip';
